[refactor] extract intersectionObserver and Gallery
