+++ src/pages/Explore.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef, useContext } from "react";
+import { useState, useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
 import { Firebase } from "../utils/firebase";
 import {
@@ -11,17 +11,10 @@ import {
 } from "@chakra-ui/react";
 import { Search2Icon } from "@chakra-ui/icons";
 import { Select } from "chakra-react-select";
-import Gallery from "../components/Gallery";
-import { AuthContext } from "../components/auth/Auth";
+import IntersectionGallery from "../components/customHook/IntersectionGallery";
 
 export default function Explore() {
-  const [user, loading, error] = useContext(AuthContext);
-  const [exploreworks, setExploreworks] = useState([]);
   const [alltags, setAlltags] = useState([]);
-  const [isShown, setIsShown] = useState([]);
-  const endofPageRef = useRef();
-  const pagingRef = useRef(null);
-  let isFetching = false;
   const [selectedOption, setSelectedOption] = useState(null);
 
   const [searchParams, setSearchParams] = useSearchParams();
@@ -33,40 +26,12 @@ export default function Explore() {
     });
   }, []);
 
-  useEffect(() => {
-    const pagingObserver = new IntersectionObserver((entries) => {
-      if (entries[0].intersectionRatio <= 0) return;
-      if (isFetching) return;
-      if (typeof pagingRef.current === "undefined") return;
-      isFetching = true;
-
-      Firebase.getWorks(pagingRef.current, queryTerm).then(
-        ({ fetchWorks, lastVisibleWork }) => {
-          setExploreworks((pre) => [...pre, ...fetchWorks]);
-          setIsShown((pre) => [
-            ...pre,
-            ...Array(fetchWorks.length).fill(false),
-          ]);
-          pagingRef.current = lastVisibleWork;
-          isFetching = false;
-        }
-      );
-    });
-    pagingObserver.observe(endofPageRef.current);
-    return () => {
-      endofPageRef.current && pagingObserver.unobserve(endofPageRef.current);
-    };
-  }, [queryTerm]);
-
   function handleSubmit(e) {
     e.preventDefault();
     let formData = new FormData(e.target);
     let newQuery = formData.get("query");
     if (!newQuery) return;
-    pagingRef.current = null;
     setSearchParams({ query: newQuery });
-    setExploreworks([]);
-    setIsShown([]);
   }
   return (
     <Box mt={12}>
@@ -110,13 +75,8 @@ export default function Explore() {
         </form>
       </Stack>
       <Stack mt={12}>
-        <Gallery
-          works={exploreworks}
-          isShown={isShown}
-          setIsShown={setIsShown}
-        />
+        <IntersectionGallery term={queryTerm} currentUserID={null} />
       </Stack>
-      <div ref={endofPageRef}></div>
     </Box>
   );
 }
